Move Trust_system out of Coda_base
The old plan was that Peer_trust would be a library independent of any
Coda-specific stuff, and that Coda_base.Trust_system would contain
the instantiated trust system. This creates circular dependencies though,
Coda_base has dependencies which need to record misbehavior. So I moved the
Trust_system module into what used to be the peer_trust library and renamed
that to Trust_system.