Add 5 proposer test (#1870)
* tweak five-proposer test

* rewrite prefix check to query transition frontiers

this adds best_path, and fixes a bug in path_map when the
breadcrumb you start on is the root already.

* make account balances larger so that the payment test doesn't influence the stake distribution much

* formatting

* try and get a useful backtrace on that

* don't try updating chain for 'offline' nodes

* coda_worker_testnet: add synchronization around node shutdowns

this prevents doing restarts etc while there is an RPC currently happening.
I believe my ad-hoc synchronization will never deadlock, but a careful review
would be highly appreciated

* make best_path part of Api like everything else

* reformat

* during teardown, wait for ongoing RPCs to complete

* multiproposer test: support sending some payments

* coda_worker_testnet: make workers an array, update it on start

previously, when we would restart a node we would still be using
the old connection object.

* don't do payments for now

* remove backtrace recording from coda_bootstrap_test

* incr/decr