Syncable ledger: fetch both children of a node at once + delete validity tree (#1984)
* Syncable ledger: fetch both children of a node at once + delete validity tree

We fetch both children at once, from the same peer, so that we can be sure
whether or not the peer who sent us the hashes lied. In the previous
implementation, where we fetched them separately from different peers, all we
knew what that at least one lied. This will make the upcoming punishment
changes more reliable.

I deleted the validity tree because it's complicated, mostly uncommented, has
bugs and is really only useful if we are deferring updating the hashes in /
writing to underlying ledger until syncing is complete. Which we aren't. I
expect we'll end up doing that and putting the validity tree or something like
it back in at some point in the future.

* Remove unused module signature

* Remove support for in-memory Merkle ledger

* Catch errors better, improve log message