Fix transaction pool test: don't check impossible situation (#2028)
@ghost-not-in-the-shell modified the test to check behavior when switching
forks in situations where a transaction is in both forks. What he wrote tests
an impossible situation though: a transaction that is already included in the
best tip will never be in the pool. Either it was received from gossip after
the block including it was processed, in which case the nonce check would fail,
or it was received before, in which case it would be removed by the
best_tip_diff handler.

This commit tests the behavior in a situation that can actually happen.