logproc: hashfarm
this uses String.hash to select from a generated list of
about a million names. pass --farm to logproc and it will
turn any Data_hash into an adjective color animal.

this uses String.hash to select from a generated list of about a million
names. pass `--farm` to logproc and it will turn any Data_hash into an
adjective color animal. or use `--name-value
428799327871634140614883825262785951378449376286542231472693588329549637135276814515289687`
to figure out what name a particular state hash was given.

I stole this idea (and the wordlists) from @jkrauska.

Some nits I dislike about the implementation:

1. Collisions are fatal. I did this for consistency, so that when a
collision does happen, at least you'll know about it if it's from the
same logstream. The main alternative I see to failing is to return the
original string.
2. The JSON you end up querying is different from the JSON written to
stdin. This is because I erase the `logproc_interp` thing into either a
name or the actual data hash. This might be confusing, but I think it's
less confusing than having to know about the underlying data_hash
encoding.
3. The command line parsing is a bit gross, pretty much the easiest thing I could think of.

Concretely, this shrinks a hash from 90 characters to, on average, 21.
And it gives them pronounceable names!