O(1) section redesign (#2129)
* O(1) section redesign

Section is re-done in the new technical style. Not done are the
equal-sign borders (just a placeholder for now). Looks good on my
iPhone.

The fonts are encrypted inside a zip file to mitigate the
potential pirating of these fonts, ask @bkase for the password if
you have creds to see the font files. Decrypt by running
`npm run decrypt`.

* Delete unused investors

* Attempt to fix CI

* Npm install before decrypt

* Pragmata is in the website folder