Refactor App logic a bit (#2185)
Tray and window logic was quite similar and not necessary to be inside
of App. Moved it out.

Single is a singleton-like holder (but not quite a singleton) that can
also drop it's cached item. The specific Tray and Window modules we need
have an App prefix because BsElectron will shadow Tray and Window
modules.