Monads and File I/O on JSON "db"
Built out a monad interface in the Jane Street style along with Result,
Task, and Task.Result.

Task is a lazy async primitive that when "fork"ed is evaluated.

Result is result. The idea with including a custom Result here (and
custom monad signature) is to keep us easily adaptable should we decide
to back Belt or Tablecloth or Base.

For now, we can use infix monad operators until we set up the
`let%bind` ppx.

The FlatFileDb stores hello-world then loads it to ensure that the task,
result, monad infrastructure actually works.