Support for modals and deeplinking
Typesafe routes can be sent across the process boundary via a
`Deep_link` message.

Currently all deep links just open different modals. The modals are
simple absolutely positioned divs that sit on top of the whole
application.

Opening the modals from within the webapp works. Opening the modals from
within the tray while the webapp window is open works. Opening the
modals from the tray while the webapp window is closed works.

The MainCommunication and RendererCommunication files must be separate
modules, one for the renderer process (this consumes main communication
to communicate with the main process), and one for the main process
(this consumes renderer communication to communicate with the render
process).

Styling intentionally unimplemented for now.