Introduce settings communication (#2243)
* Introduce settings communication

Add a CallTable to make it possible to `Task` across the IPC barrier
and used that to send/receive settings changes. For now, settings is
loaded just once from the tray, but can be updated by the Renderer as
many times as necessary.

Bs-json introduced for managing JSON encode/decode temporarily until we
can get something more automated (be it a ppx or some other sort of
preprocessor like quicktype).

Routes include the settings payload as a query parameter because rather
than being sent it as a separate message because we _need_ settings
information in order to do even the first render of the UI. For example,
imagine there is theming information in there.

TODO: Make a custom react hook for the settings message
TODO: Introduce testing infrastructure and then write tests for this

* Requested changes

* Added tests

* Free memory when call is resolved