Add content to Send/Delete modals
Content in place for Send/Delete modals. Cancel button correctly closes
the modals and clicking outside the content window also closes.

ModalButtons are reused between the two modal states and it seems like
this is the only area with that style Button is used, so I inlined the
"Button" definition in the ModalButtons module.