Settings reactively updates across processes
CallTable now is typed with GADTs and handles the settings changes by
the tray process responding with the new settings.

Luckily we only need to update the settings after the application starts
via changes initiated by the render process. Whenever this is no longer
true we'll need to handle the extra messages and correctly call
`setSettingsOrError`.