Added more logging to tf persistence (#2281)
* Added more logging to tf persistence

It is quite difficult to identify the cause of this flaking test in tf
peristence

I added more logging so that I can get detailed information of why the
test fails. Namely, the worker will log the mutants that they produce.
Also, more information is printed when encountering Update_root diff.

I have seen the failure come up twice out of a 1000 rounds. The failure
happens after encountering an Update_root diff.
It seems like the state_hash of the Update_root mutant matches with the
ground truth.  So, it might be that other root data information,
pending_coinbase and scan_state, are not being updated correctly.

* Fixed Simple Nit