Implement wallet management w/ graphql
There exists wallet management logic on the daemon now. This is built on
top of the secret-file infrastructure (for now with a null password).
Due to cyclic deps, the secret-file related files were moved from
cli_lib to secrets along with the new wallet management logic.

GraphQL support was tested via inspection with the `graphqurl` tool.
Wallet logic was tested with unit tests. I also verified that executing
an `addWallet` mutation and then a `ownedWallets` query fetched the
public key of newly created wallet.