Fix 2 race conditions with proposer and transaction pool (#2352)
These were causing the "Invalid user command! Error was: Nonce in account 1
different from nonce in transaction 0" errors in fake_hash:full-test.

In proposer.ml we were getting the current best tip to propose off of it, then
waiting for an IVar, then getting the transaction pool. If the best tip changed
while we were waiting, we would see transactions that were valid against the new
best tip and potentially not the one we basing our proposal on. I changed it to
get the transaction pool at the same time as the best tip, before waiting.

We were also waiting for the proposed transition to be accepted into the
transition frontier before starting to propose again. The second race condition
is when our proposed transition has been accepted and we start to propose again
before the transaction pool is updated to remove now-invalid transactions. I
fixed by changing it to update the diffs (and consequently the transaction
pool) before calling Transition_registry.notify, which wakes up the waiting
proposer. This way the transaction pool is guaranteed to be updated in time.