Model coda cli settings properly
Here precisely captures the state necessary for keeping track of which
flags to pass to the CLI and if we need to download any keys.

This will later be consumed by logic that maps this data to actions
which when consumed via some reducer will turn on and off the CLI
client.

Verified manually with figma and hand-drawn diagrams that we've captured
all edge cases.

Co-authored-by: Pranay <o1pranay@users.noreply.github.com>