Add application-side logic for managing coda.exe (#2521)
* Add application-side logic for managing coda.exe

Moved types around and functored over side-effectful things in the
application state so that we can unit test the reducer effectively.

The reducer supports starting coda more than once with the same args
while it's in the middle of starting or if it's already started. If one
  tries to start it with different args, the process it killed before
  being started again.

Window state is moved over to the application reducer as the reducer
needs to know about the window in order to respond to coda.exe events
properly.

State machine is partially explored with unit tests, but is not
exhaustively tested.

* Fixup conflicts, PutWindow

* No ready + supports process death

Rip out graphql ready as that will be handled from the render side.
Process death is captured and sent over IPC.

Unit tests also support process death now.

* Update frontend/wallet/src/common/CallTable.re

Co-Authored-By: Avery Morin <avery.schmavery@gmail.com>