Uses Coda.exe for certain GraphQL commands
addWallet / getWallets are both piped through to coda.exe instead of the
graphQL faker. Coda.exe is started at the beginning of the render
process starting up via the `<CodaProcess>` component.

We were using hooks wrong in two different cases:
1. `useEffect` runs every render call by default, `useEffect1` and
passing `[||]` causes it to only run once (you can also pass dependences
to that array to get it to run whenever the deps change)
2. `useReducer` in hooks must be pure. Side-effects need to be handled
in useEffect blocks. See `Hooks.Reducer.useReducer` for a useReducer
that can handle effects.
Note: The Application's reducer still does use side-effects. We may want
to change this in the future to mirror the behavior of the react one
more closely.

This is hacked into SettingsPage via passing `dispatch` down manually.
It doesn't seem right to stick this in context, but we can figure out
where this should go later. SettingsPage has a fake button to trigger a
restart of the process with different arguments.

Currently, the only thing stopping this from working properly as far as
I can tell is the weird behavior with `kill` vs `ctrl-c` in bash.

A few new unit tests were added, and also tested by playing with the
app.