Support changing stakers at runtime  (#2567)
* Support changing stakers at runtime

We always start proposing now and with a set of keypairs. If the set is
empty it's effectively the same thing as not proposing. Proposing is not
interrupted by a change in staking keypairs, but it is woken up by this
change. If we are in the middle of staking with different keypairs, that
action is finished before the new keys kick in.

Local_state is now a table from public keys to delegation info and we
check vrfs for all the keypairs that we wish to stake.

I've included a handler in Ledger.ml as it may be useful later even
though I ended up not using it in this PR.

Haven't tested thoroughly yet.

* Requested changes

1. Precompute compressed public keys via Keypair.And_compressed_pk.Set.t
2. Move Agent to Otp_lib and make it backed by a ref
3. Fix N+1 query problem on delegatees
4. Store unique epoch/slot pairs for each pk we're tracking and set
intersect with the new keys

* Requested changes

* Flipped a bool, clean up code

* Off-by-one error in sparse_ledger.iteri

Plus quickcheck test for confirmation

* Remove extra debug logging

* Safely lookup delegators

* Requested changes on the agent w.r.t. readonly

* Merge fixup