Support daemonStatus in GraphQL
Derived using `Status.Fields` so we'll get a compile error if we forget
to update the GraphQL schema when we change the record.

Built out a little toolkit of field-based reflection functions to easily
stitch together graphql representations of our fields.

Added `deriving fields` to a few type dependencies of status record.