Simplify snark work spec - remove statement type parameter
This required defunctoring a ton of stuff, and I had to rewrite some tests since
mockery isn't possible for the defunctored stuff anyone. The new tests for
staged ledger are much slower since they now involve an actual ledger and have
to do Pedersen hashing, but they're much more complete. Due to some refactoring,
the txpool tests are also slower, but I think it's worth it to make things
simpler and more realistic.