CLI version info (#2949)
The version command to the CLI executable was printing garbage info, because the Jane St Command processor handles that one specially. To use it properly requires generating a C file from a script that's apparently not in the Jane St libraries.

Instead, we intercept version (and version -help) on the command line, and print the information desired.

There was already a script to generate the git commit as a value in an .ml file, here extended to generate the git branch. So now we have:

```
$ coda version
Commit 087220d on branch release/beta
```

The -help message mimics what the Jane Street library would have printed:

print version information

```
  coda.exe version

=== flags ===

  [-help]  print this help text and exit
           (alias: -?)
```

Closes #502.