Properly unregister masks when moving tf root + better error me… (#2943)
Prior to this commit we unregistered the masks of the children of the old root
that were not on the winning chain, but ignored any descendants of those
children. This commit correctly unregisters all the masks. I also made the error
messages when you try to remove the parent of a mask that isn't attached
properly more informative.

This won't fix the issue @figitaki was seeing, but it'll at least make debugging easier since there will be less garbage in the visualizations.