Update *all* children in remove_and_reparent_exn
Prior to this, you passed the children as a parameter to the function. If you
missed any, you'd put the mask system into an invalid state. This happened in
move_root in the transition frontier, because there may be masks parented off
a ledger in the TF that aren't part of breadcrumbs in the TF, for various
reasons. The new code looks them up in the register, preventing this kind of
error.