Fix/unregistered mask testnet (#3009)
* Assert that we aren't orphaning any masks by unregistering their parents

* Update *all* children in remove_and_reparent_exn

Prior to this, you passed the children as a parameter to the function. If you
missed any, you'd put the mask system into an invalid state. This happened in
move_root in the transition frontier, because there may be masks parented off
a ledger in the TF that aren't part of breadcrumbs in the TF, for various
reasons. The new code looks them up in the register, preventing this kind of
error.

* Fix unregistering masks in with_ephemeral_ledger

* a little cleanup from PR comments