Fix fd leak when connect() fails (#3056)
* Fix fd leak when connect() fails

We were leaking the socket when connect() failed, since our logic for closing
them uses Reader.close, and no reader is created until after we have a
connection. We had our own code to set up sockets bound to a particular local
IP for outbound connections, but it turns out you can create a
Where_to_listen.t that includes the local IP. I changed it to do that, and to
pass the Where_to_listen.t to Rpc.Connection.with_client, which handles setting
up the sockets and disposing of the resources properly for us.

We were also calling Reader.close and Writer.close manually when handling error
conditions. This is redundant inside of a Tcp.Server.create callback, the
socket is closed automatically when the Deferred.t returned by the callback is
determined. Removed those calls.

* Remove out of date comment