An unsafe account import (#3134)
* An unsafe account import

A small step towards a new account CLI model. This step conveniently
provides an easy way to inject private key files into the database that
the GraphQL API uses to think about accounts.

By design, it does not need a running daemon to execute, and as such it
requires you pass the conf_dir if it is in a non-standard location.

* Requested changes