HACK: non-fatal proposer timeout (#3168)
Closes #3160

We're seeing large (largest I saw so far was ~17s, @ghost-not-in-the-shell has seen more) async cycles causing this timeout to trigger. I turned it up, so hopefully it won't happen as often on decent CPUs, but also remove the exception. If the timeout fires, we aren't guaranteed to make our next proposal on top of our own block, even if it ended up displacing the best tip. I think this actually can break Ouroboros, and is insecure, but shouldn't be prone to accidental forks with the reasonable k that @nholland94 is adding. I think. We should disable this at some point, which I've created #3167 to track.