CODEOWNERS simplification for faster review cycles
If code that you owned, was _so clean_ that we didn't think it needed to
be blocked by your review, we removed that line. Congrats!

For the rest of the code, there is now a massive reduction in
bottlenecks on reviews needed for a given PR. Many more core
contributors were added to exisiting libs especially ones with few
people.

Unfortunately, there are still a few libraries that only a single person
can review. We'll hopefully get better in the future to not need that.

If you feel like we removed a library that you believe is too complex
and does still need review from specific people, please leave a note.
Also, if you're uncomfortable reviewing something that you've been
added to mark as such on the PR.

Co-authored-by: Nathan Holland <nholland94@gmail.com>