Eager bootstrap connects to multiple people (#3251)
Before a node does eager bootstrap, it will now wait until it is connected to a good amount of peers (Namely 8) or timeout. Then, it will ask all the other peers about their best tip. The node should sync with the node with the peer who has the best block,

I created a unit test that verifies that this thing works.