Add optional libp2p discovery (#3323)
use -libp2p-discovery as daemon arg to enable. There is a -libp2p-port argument but it is IGNORED at the moment in favor of the hardcoded port 28675 (arbitrary). When we remove the Haskell Kademlia, we'll take over the existing port infrastructure.

There's a -disable-old-discovery that turns off old discovery as well. Risky!

* cli flags

* Gossip_net: keep the whole Config.t

* add libp2p_membership field

* missing opam file

* start threading it all through

* big ole nasty commit that does a bunch

* libp2p keypair + initial peers

* inject new libp2p peers back into haskell membership