Refactor SNARK bundles to use One_or_two library
Transaction SNARKs and associated data always come in bundles of one or two.
Prior to this, we either used a list or a pair like 'a * 'a option to express
this. I've replaced all of those with One_or_two.t, which buys us type safety,
self documenting code, and saves us ~60 lines of code due to reducing
redundancy.