Script to enforce ppx_coda always used (#3646)
The version syntax linter in #3633 only works if `ppx_coda` is used to preprocess a file.

The script in this PR is added to the `lint` step of CI to make sure that every `dune` file contains `ppx_coda` in its list of preprocessors, except for those in `src/external`, `src/lib/snarky`, and, of course, in `src/lib/ppx_coda` (and the `src/_build` directory, if you've already built).

The existing `dune` files have been updated to include `ppx_coda` where they did not already do 
so.

Update: I'll need to build the docker image to install the Python sexplib library. Done.