Checks s3 before failing cached read of pvkeys
To make it easier to reuse pvkeys, `Cached` now consults an http
endpoint before giving up and trying to autogenerate pvkeys.

We also exit the build with 0xc1 only on CI builds (so we may push the
keys to s3 and fork the pipelines).

TODO: Find the right http GET path, and test this code.