Checks s3 before failing cached read of pvkeys (#3845)
* Checks s3 before failing cached read of pvkeys

To make it easier to reuse pvkeys, `Cached` now consults an http
endpoint before giving up and trying to autogenerate pvkeys.

We also exit the build with 0xc1 only on CI builds (so we may push the
keys to s3 and fork the pipelines).

TODO: Find the right http GET path, and test this code.

* Requested changes

* Circle yaml changes

* Run reformat

* Properly print paths

* upload keys

* public-read

* More detailed error message + conduit-async for ssl

* Minor fixes, only regen in testnet profile

* add ls

* add sleep

* iterate uris

* Use curl instead

* Better curl; more logs

* Mkdir~p s3_install_path

* Up autogen

* remove workspace, search tmp s3 for keys for deb

* remove order

* remove kadem, cosmetic s3

* fix path

* force /tmp

* artifactory not being used for mac