Transaction expiry (#3963)
This PR adds a `valid_until : Global_slot.t` field to user commands. Transactions are valid if the current slot is `<=` this value.

The default behavior (if a user does not specify a `valid_until` slot) is to set `valid_until` to the maximum uint32 so that the transaction never expires.

This feature actually checking against the correct "current slot" depends on some changes @psteckler is making to thread through the current protocol state to `Transaction_logic.apply_user_command_unchecked` and `Transaction_snark.apply_tagged_transaction`. As of now I've stubbed this part out.