Website: Refactor docs metadata (#4211)
Removes the Metadata component in favour of passing in the title directly to the page component. This lets us only set the title and other page metadata in one place.

This also unblocks making the title prop on Page required, so this does that too.

Also sets a description on the genesis page.