best tip diff extension fix (#4418)
Transactions from the new best tip was being added twice to the diff. When trying to remove a transaction from the pool the second time, it wouldn't be in the pool and hence the nonce mismatch error. Added some extra information to the error logs in the case #4298 is still seen

Also added a check for the account creation fee before adding txns to the pool. If the receiver account of a txn is new and the txn amount is less than the account creation fee then it won't get added to a block.

The response of the `send-payment` command doesn't show any of the transaction pool errors (#3457) so made errors from local transactions be logged with level `Error`