Check for consensus constant overflows in the snark (#4783)
Previously we were asserting that these overflows could not happen at
snark compile time. As we move more of these constants to runtime, more
of them will be variables, and anything more than a single
multiplication (u32 * u32 -> u64) will cause the compile-time check to
fail as a potential overflow.

By specifying the number of bits that a Snarky_integer.Integer.t should
be unpacked as, this is converted to a constraint system check: if the
result is not representable in the given number of bits, a proof cannot
be built.