add a global_metadata to logger (#4832)
* add a global_metadata to logger

this is used for including the peer_id and other network identifying
info in every log message. since we don't know our peer_id until it
is generated, we can't include it in the metadata when we create the
logger. a potential alternative would be making the coda_lib.logger
mutable, but that doesn't work because loggers that extend it won't
get updated when we update that logger. instead of invasively
replacing the logger internals with some sort of mutable linked list
to parents, use a small bit of global mutable state (ugh).

* don't expose the global_metadata directly

Co-authored-by: Conner Swann <me@connerswann.me>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Nathan Holland <nholland94@gmail.com>