Carry fee excesses for 2 tokens in each transition
This change allows us to accept fees in multiple tokens. The intention
is that transactions are ordered by the token in which their fees are
paid, and the balance for each token is settled by a fee transfer before
moving on to transactions paid in the next. For example:
```
[Pay Tok1] [Pay Tok1] [Pay Tok1] [Pay Tok1] [Fees Tok1] [Pay Tok2] [Fees Tok2] [Pay Tok3] [Pay Tok3] [Pay Tok3] [Fees Tok3] ...
    [Merge Tok1]          [Merge Tok1]        [Merge Tok1, Tok2]     [Merge Tok2, Tok3]       [Merge Tok3]        ...
               [Merge Tok1]                               [Merge Tok1, Tok3]                             ...
                                     [Merge Tok1, Tok3]                                 ...
```

Under this model, we need only consider the fee excess for the fee
tokens used in the first and last transactions within a transition, and
we can check within the snark that the intermediate tokens are resolved
when merging.

This includes some additional logic to 'erase' token IDs -- replacing
them with the default token -- when the balance is settled for that
token. This ensures that we do not need to know the specific tokens
involved when building a blockchain proof from a merge proof: we can
pass the default token and zero excess in both of their places,
equivalent to the previous condition of [fee_excess = 0].