Fixes cache-through process (#5179)
Cache-through was completely broken. This PR cleans up the abstraction, and fixes everything. The newest Coda container needs the opam dir from the cache in order to run properly, so we know it works now.

The `OpamInit` module now has a nice way to run extra commands after pulling opam from the cache (if it's a hit)