Ci-build-me bot works, fixes #4988 (#5231)
Created a bot to fix #4988 

See [README](frontend/ci-build-me/README.md) for details on how to test and deploy.

Tested by applying the `ci-build-me` label to a PR, observing a buildkite job starts. Also, tried pushing to a PR with the label on it and a buildkite job starts. Importantly, pushing to a PR without the label does not trigger a buildkite job. I manually verified that this only works for PRs opened from internal PRs. See https://github.com/CodaProtocol/coda/pull/5233 .

Additionally, a comment of exactly `!ci-build-me` on a pull-request, if that comment is authored by a user that is part of the CodaProtocol organization will also trigger a build. I verified that folks who are not part of the organization, cannot trigger builds with this special comment as well (see the comment via @mobileink ).

The secrets needed to test and deploy this bot are stored in AWS secrets manager under `coda/ci-build-me-bot`.

This bot is currently live on Google Cloud Functions.