[Tokens] Simplify transaction pool (#5252)
* Do not evict commands based on source balance

This is inherently racy, and the logic no longer applies when the
account differs from the fee_payer account. These failures are captured
in the snark, where the fee_payer is still charged but no other part of
the transaction is processed.

* Remove receiver checks in the transaction pool

These failures are checked in the snark and transaction logic. When they
appear, the fee is still charged to compensate the network, but the
transaction itself fails.

* Store constraint_constants in the indexed pool

This is used by tokens commands to calculate the account creation fee
(which is charged to the fee-payer).