Hacky hack: Always use nonconsensus profile in merlin for nonconsensus (#5284)
This abuses the hack that we use to make merlin work with absolute-path
includes in our fork of ppx optcomp. Merlin is started with its cwd as
the directory that the currently edited file is in, and then walks out
until it finds a file at that path relative to a parent directory.

By inserting `src/config.mlh` paths (aliased to the nonconsensus
profile) below the `src` directory for code that only gets compiled with
the nonconsensus curves, merlin is tricked into always using the
nonconsensus profile for them, regardless of the current build profile
in `src/config.mlh`.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>