Adds support for triggering circleci builds too
In an effort to curb CI spending, we wish to change circle builds to
opt-in. You must either:

a) Add the `ci-build-me` label to your PR (this only works if you are
PRing _not_ from a fork for security reasons). A build will trigger the
moment you add the label and every time commits are pushed to that
branch.

b) Add the comment `!ci-build-me` to your PR. This will trigger one and
only one build at that moment.