Initial pass at Rosetta test-agent
The test agent process runs alongside the network + archive + rosetta.
It pokes Coda to manipulate it's state (starting/stopping staking,
sending transactions, etc). And it peeks at Rosetta to observe that
after poking the state of the world makes sense.

This PR sets up the initial scaffolding for the above system with a
simple GraphQL mutation and Rosetta request.

For now, peeking at Rosetta is done via making HTTP requests, but we can
complement it with peeking actions performed by the [Rosetta
CLI](https://github.com/coinbase/rosetta-cli) in future commits.

This agent can be used with a local demo build, but also is reusable
against QA-nets and real networks.