Rosetta agent test assures block matches mempool
After checking the mempool, stake again, and then wait for a block.

Moved `similar` check to a validation rather than a result to gather all
the errors; this lets us search through operations to find the most
similar when reporting errors. This aids in debugging.

This test uncovered a bug in the sql query mixing up source and
receiver in a payment. It's probably worth fleshing out all the cases
here.

The agent stops successfully, and the test passes.