Write transition to bootstrap pipe if the transition is for bootstrap
This fixes a 'write to closed pipe' error where we were writing to
`transition_writer_ref`, then mutating it and `kill`ing all of the pipes
that the `Transition_frontier_controller` writes to when it goes to
handle the message.

The underlying error is that we were (correctly) closing the transition
frontier controller and opening a bootstrap controller, but
(incorrectly) sending the message to the writer pipe before updating it
to point to the bootstrap controller. This behaviour matches the
original version from pre-#3974.