fix system log being flooded with messages
This change will include the new file I created at ../scripts/mina.conf in the build process
This will stop the system journal from defaulting to output to syslog. 
I don't believe this is useful as log files are already created in .coda-config and the system log is more for the system than services with large amounts of output.