feat: Changed verify function signatures to use bool
Initially, there was a bug with a few of the verification functions in
client_sdk.ml where the return value was a ocaml bool but when mapped to
javascript, it was interpreted as a number. This has been fixed and so
we can use the correct return type of a bool.