Helm: add persistent storage for block-producer
The current helm chart seems geared towards integration testing, but
for a production node operator to be able to use helm, it is desirable for the
state to persist across restarts.

Adding an option to helm chart `.coda.persistentStorage` achieves this.

Defaults to false.

Sumbitted as part of my attempt to make https://github.com/midl-dev/mina-pulumi
project work well with default chart.