helm: relax requests for block producer to 7 vcpus instead of 8
The mina documentation recommends a c5.2xlarge instance to run the
block producer, but using such an instance in EKS will result in the
pod not starting due to not enough CPU.

Changing the request to 7 leaves enough CPU for the kubernetes system
daemons. It should be minimally disruptive to other consumers of this
chart.

Sumbitted as part of my attempt to make https://github.com/midl-dev/mina-pulumi
project work well with default chart.