fix: Moved SPREADSHEET_ID to helm chart from Dockerfile
Moved the env var for SPREADSHEET_ID from the Dockerfile to the helm
chart. It was a little awkward that there was one env var being exported
in the Dockerfile and one being exported in the helm chart. Made an
effort to combine them into a single source of truth.