SNAPP Electron App
This app is a demo that uses snapps to prove that a user has a credit
score above a certain threshold. It shows a form to the user and has
fields for the creditkarma service credientails. Once the user submits
their creds, it will attempt to log the user in and save their credit
score locally. If it is a above a certain threshold, it will generate a
snapp proof that is used for confirmation by a 3rd party that the user
indeed has a credit score above the threshold.