Fixed error with credit score below threshold
The check that does the credit score threshold was not exiting the
function if it was below the threshold. Made a fix for that by adding a
return statement if the check fails. Additionally fixed an error with
the error message as it wasn't initializing the outputPath variable in
scope