Added wrapper for signedRosettaTransactionToSignedCommand
Added a JS wrapper for the `signedRosettaTransactionToSignedCommand`
that converts a signed Rosetta transaction to a JSON string
representation of a Signed_command. The usage of this function is
intended for converting Rosetta transactions to a format that GraphQL is
happy with.

Additionally added some property based testing with a mock Rosetta
transaction.