Remove support for AppImage in package.json
* The Snapp binary that gets packaged needs a system dependency for
Linux which is `libjemalloc`. For users that do not have this installed,
the Snapp binary isn't able to be called and thus defeats the purpose of
the Electron app. Instead, we can specify a build depenedency for .deb
packages and include `libjemalloc`. This way it is installed on the
users system and can use the app as normal. This also means that the
AppImage format can no longer be supported without some further work
arounds.