Fix Mac OS Snapp proof output
When calling upon the snapp executable on MacOS, the Electron app
continously states that the proof generation failed. This is an issue
with exec() as it checks the error code of the last command to be
0 and issues an error otherwise. On Mac, we are getting an error code
of 1 even though the output is successfully being generated. Added
a check for Mac to continue with writing the proof if this is the
case.