Explicitly install the mina-dev-noprovingkeys package
Apparently we need to do this: apt-get will always attempt to install
the latest version of all dependencies, and then fails if these latest
versions do not match the constraints. Unfortunately for us, the
'latest' version is the largest version string lexicographically, which
causes us to arbitrarily prefer some commits over others depending on
which has the smallest hash.

A partial workaround would be to include the 'height' of the commit
before the commit hash itself, but this will still be brittle against
rebases etc.

This shouldn't be an issue for official releases: the version number
increase should ensure that the latest version is the correct one.