Added stagedLedgerProof to block object
Implemented the `stagedLedgerProof` endpoint on the block object since
we need access to a block hash to search in the transition frontier.
Currently, this is implemented by finding a breadcrumb associated to the
queried block and checking if the stage_ledger_hash matches the one in
the current block object. This is experimental right now and is subject
to changing.