[ocaml] instruction to set up local switch
To build outside of docker a developer needs to create an [opam switch](https://opam.ocaml.org/doc/man/opam-switch.html) from a custom `ocaml.export` file for reasons.
This switch is then implied, and the developer needs to remember to switch back to default, or to whatever they prefer, when working on other projects.

This isn't a great experience (which I ran into when trying to install a dependency that required a greater ocaml version in another project).

This PR proposes a [local switch](https://opam.ocaml.org/blog/opam-20-tips/#How-local-switches-are-handled) to transparently use the mina switch when in the mina repository.