Responding to feedback on logging section
A good point was made about restructuring the proposed logging
architecture. Instead of polling on an interval with Docker CLI which
can lead to latency issues, we can redirect all stdout to a pipe shared
with the test executive. This change reflects the following changes in
the rfc to address that feedback.