Add implementation for network_config and stubbed interfaces
This adds the functionality to create a base docker-compose file from
the specified test configs in the integration framework. Each node will
be assigned to a service that will be running a container for each node
in the network.

This implementation currently supports creating a docker-compose file,
creating a docker stack to run each container, and basic monitoring
before continuing the rest of the test_executive execution.

Lots of code in this PR has been stubbed and copied to make the compiler
happy. Much of this code will be refactored to support the changes
needed for the local engine later.