Added Archive node support to local test engine
Added support for archive nodes in the local engine. If archive nodes
are specified in the test plan, the test executive will download the
archive schema to the working directory to use the schema as a bind
mount for the postgres container. The archive node will then connect to
the postgres container as usual.