Added GraphQL networking to docker containers
Added functionality for GraphQL networking to the docker containers.
This work was mostly just adding the '--insecure-rest-server' flag to
the nodes and opening up the corresponding rest port within the
docker-compose file.

To allow communication to the docker network from localhost, we map
ports on the host starting at 7000 to port 3085 to the node which is
used for GraphQL communication.