feat: Basic implementation of http-snapp-server
Added a basic implementation of the http-snapp-server. This is meant to
work in conjunction with the `httpsnapp` library which will proxy
requests from the user to this server.

The server will make a request on behalf of the proxy and sign the
message with the Mina client-sdk. If there is an error, we simply
propogate it back to the `httpsnapp` caller.