feat: Use axios instead of node-fetch and pass in URI prefix
Two main changes are included in this commit. Firstly, we switched out
node-fetch for axios as we want this code to be able to work both in the
browser and on a server. Secondly, we applied a currying technique to
allow the caller of `verify` to pass in a URI prefix for the oracle
endpoint instead of having a hardcoded value for now.