feat: Include request as raw bytes in response
Added functionality to include the request object as raw bytes in the
returned response to the user of this API. Currently, we are not
including the request data as part of the signed data as our client sdk
cannot support it. Instead, we just mock the data we expect to be
returned and go on with our day, for now.