Rosetta: Ensure duplicate error fires
The daemon can't disambiguate between duplicates and bad nonces on
transactions. So we'll use a small ring buffer as a cache of
successfully sent transactions. When we receive a bad nonce error, we'll
check the cache to see if we've already sent this transaction this
session.

Tested by running a build connected to one of the nodes on devnet and
confirmed that the duplicate error can be seen.