Rosetta: Duplicate transactions hash preimage
The hash in the database is not reproducible (sadly) from the rosetta
representation of the transaction at the point in the construction
process. As such, we can't rely on it for duplicate tracking. Instead
we'll just check all the values. Since we have the nonce, these values
should be sufficient for the check.