Pubsub: support two layers of topics
Problem: to deliver switch to new topics as a soft-fork,
we need a mechanism that would allow nodes
to decide on set of topics to communicate on based on
versions of one other.

Solution: use WithSubcriptionFilter of Pubsub configuration
to filter out old topics when topics corresponding to the new version
are present.