Use hash-based message id function
Problem: default message id function used for pubsub is
based on public key of the message creator and some nonce.
Hence, it is possible for the same message to appear twice on the same
topic provided it was sent by different creators.
This might be problematic for some types of gossip, especially in the
light of forecoming softfork introducing a versioning mechanism for
pubsub topics.

Solution: use hash-based message id function.