Eclipse test
Problem: it might be possible for a new node (Carol) not to be able to
connect to a network due to seed node (Bob) having far too many
connections.

Solution: implement a test which checks this behavior.
Currently, TODOs are left in test. The point to the fact that when
implemented properly, test would fail. At a later point we
must change the behavior and satisfy the TODOs.